refactor: remove empty init.lua
